% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_predict.R
\name{.emulate_pctile}
\alias{.emulate_pctile}
\title{Compute percentile matching Stata's ceiling-index method}
\usage{
.emulate_pctile(x, pct)
}
\description{
Compute percentile matching Stata's ceiling-index method
}
\keyword{internal}
