% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_splines.R
\name{.emulate_rcs_basis}
\alias{.emulate_rcs_basis}
\title{Compute Harrell RCS basis at given x values with given knots}
\usage{
.emulate_rcs_basis(x, knots, df)
}
\arguments{
\item{x}{Numeric vector of values}

\item{knots}{Numeric vector of knot positions (length = df + 1).
Includes boundary knots at min/max.}

\item{df}{Degrees of freedom}
}
\value{
Matrix with df columns (basis variables)
}
\description{
Compute Harrell RCS basis at given x values with given knots
}
\keyword{internal}
