% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_weights.R
\name{.weight_censor_arm}
\alias{.weight_censor_arm}
\title{Fit censor weight model for one arm (stratified)}
\usage{
.weight_censor_arm(
  dt,
  id_col,
  censor_col,
  arm_col,
  arm_val,
  fu_col,
  trial_col,
  d_cov,
  n_cov,
  weight_col,
  quiet = FALSE
)
}
\description{
Fit censor weight model for one arm (stratified)
}
\keyword{internal}
