% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_weights.R
\name{.weight_censor_pooled}
\alias{.weight_censor_pooled}
\title{Fit pooled censor weight model}
\usage{
.weight_censor_pooled(
  dt,
  id_col,
  censor_col,
  arm_col,
  fu_col,
  trial_col,
  d_cov,
  n_cov,
  weight_col,
  quiet = FALSE
)
}
\description{
Fit pooled censor weight model
}
\keyword{internal}
