% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_weights.R
\name{.weight_switch_arm}
\alias{.weight_switch_arm}
\title{Fit switch weight model for one arm (stratified)}
\usage{
.weight_switch_arm(
  dt,
  id_col,
  treat_col,
  arm_col,
  arm_val,
  fu_col,
  trial_col,
  cens_col,
  d_cov,
  n_cov,
  weight_col,
  quiet = FALSE
)
}
\description{
Fit switch weight model for one arm (stratified)
}
\keyword{internal}
