% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_weights.R
\name{.weight_switch_pooled}
\alias{.weight_switch_pooled}
\title{Fit pooled switch weight model (across arms)}
\usage{
.weight_switch_pooled(
  dt,
  id_col,
  treat_col,
  arm_col,
  fu_col,
  trial_col,
  cens_col,
  d_cov,
  n_cov,
  weight_col,
  quiet = FALSE
)
}
\description{
Fit pooled switch weight model (across arms)
}
\keyword{internal}
